backend/lambda/requirements.txt
boto3>=1.34.0
mcp>=1.0.0
httpx>=0.25.0
anyio>=4.0.0
python-dotenv>=1.0.0